requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
PyMySQL==1.1.0
cryptography==41.0.7
gunicorn==21.2.0
python-dotenv==1.0.0
eventlet==0.33.3
requests==2.31.0
Pillow==10.1.0
anthropic==0.8.1
